Split parser import declarations
