+++ src/parser/mod.rs
@@ -16,6 +16,7 @@ mod core;
 mod declaration_types;
 mod declarations;
 mod expressions;
+mod import_declarations;
 mod patterns;
 mod precedence;
 mod statements;
